server/requirements.txt
mysql-connector-python
flask
flask-cors
Flask-Mail
Flask-Cors